+++ core/src/avm1/globals.rs
@@ -4,7 +4,6 @@ use crate::avm1::listeners::SystemListeners;
 use crate::avm1::return_value::ReturnValue;
 use crate::avm1::{Avm1, Error, Object, ScriptObject, TObject, UpdateContext, Value};
 use crate::backend::navigator::NavigationMethod;
-use crate::display_object::{DisplayObject, TDisplayObject};
 use enumset::EnumSet;
 use gc_arena::MutationContext;
 use rand::Rng;
@@ -106,228 +105,6 @@ pub fn get_nan<'gc>(
     } else {
         Ok(Value::Undefined.into())
     }
-}
-
-pub fn load_movie<'gc>(
-    avm: &mut Avm1<'gc>,
-    context: &mut UpdateContext<'_, 'gc, '_>,
-    _this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<ReturnValue<'gc>, Error> {
-    let url = args
-        .get(0)
-        .cloned()
-        .unwrap_or(Value::Undefined)
-        .coerce_to_string(avm, context)?;
-    let target_or_path = args.get(1).cloned().unwrap_or(Value::Undefined);
-    let target = if let Value::Object(target) = target_or_path {
-        let target = target.as_display_object();
-
-        if target.is_none() {
-            log::warn!("loadMovie: Object passed as target is not a movie clip");
-        }
-
-        target
-    } else {
-        let target_path = target_or_path.into_string();
-        let target_clip =
-            Avm1::resolve_dot_path_clip(avm.target_clip(), context.root, &target_path);
-
-        if target_clip.is_none() {
-            log::warn!("loadMovie: Could not resolve target path {}", target_path)
-        }
-
-        target_clip
-    };
-    let method = args.get(2).cloned().unwrap_or(Value::Undefined);
-
-    if let Some(target) = target {
-        let method = NavigationMethod::from_method_str(&method.coerce_to_string(avm, context)?);
-        let (url, opts) = avm.locals_into_request_options(context, url, method);
-        let fetch = context.navigator.fetch(url, opts);
-        let process = context.load_manager.load_movie_into_clip(
-            context.player.clone().unwrap(),
-            target,
-            fetch,
-        );
-
-        context.navigator.spawn_future(process);
-    }
-
-    Ok(Value::Undefined.into())
-}
-
-pub fn load_movie_num<'gc>(
-    avm: &mut Avm1<'gc>,
-    context: &mut UpdateContext<'_, 'gc, '_>,
-    _this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<ReturnValue<'gc>, Error> {
-    let url = args
-        .get(0)
-        .cloned()
-        .unwrap_or(Value::Undefined)
-        .coerce_to_string(avm, context)?;
-    let level_id = args
-        .get(1)
-        .cloned()
-        .unwrap_or(Value::Undefined)
-        .as_number(avm, context)? as u32;
-    let layer = avm.resolve_layer(level_id, context);
-    let method = args.get(2).cloned().unwrap_or(Value::Undefined);
-    let method = NavigationMethod::from_method_str(&method.coerce_to_string(avm, context)?);
-    let (url, opts) = avm.locals_into_request_options(context, url, method);
-    let fetch = context.navigator.fetch(url, opts);
-    let process =
-        context
-            .load_manager
-            .load_movie_into_clip(context.player.clone().unwrap(), layer, fetch);
-
-    context.navigator.spawn_future(process);
-
-    Ok(Value::Undefined.into())
-}
-
-pub fn load_variables<'gc>(
-    avm: &mut Avm1<'gc>,
-    context: &mut UpdateContext<'_, 'gc, '_>,
-    _this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<ReturnValue<'gc>, Error> {
-    let url = args
-        .get(0)
-        .cloned()
-        .unwrap_or(Value::Undefined)
-        .coerce_to_string(avm, context)?;
-    let target = args.get(1).cloned().unwrap_or(Value::Undefined);
-
-    if let Value::Object(target) = target {
-        let method = args.get(2).cloned().unwrap_or(Value::Undefined);
-        let method = NavigationMethod::from_method_str(&method.coerce_to_string(avm, context)?);
-        let (url, opts) = avm.locals_into_request_options(context, url, method);
-        let fetch = context.navigator.fetch(url, opts);
-        let process = context.load_manager.load_form_into_object(
-            context.player.clone().unwrap(),
-            target,
-            fetch,
-        );
-
-        context.navigator.spawn_future(process);
-    } else {
-        log::warn!("loadVariables: Was not called with a valid target");
-    }
-
-    Ok(Value::Undefined.into())
-}
-
-pub fn load_variables_num<'gc>(
-    avm: &mut Avm1<'gc>,
-    context: &mut UpdateContext<'_, 'gc, '_>,
-    _this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<ReturnValue<'gc>, Error> {
-    let url = args
-        .get(0)
-        .cloned()
-        .unwrap_or(Value::Undefined)
-        .coerce_to_string(avm, context)?;
-    let level_id = args
-        .get(1)
-        .cloned()
-        .unwrap_or(Value::Undefined)
-        .as_number(avm, context)? as u32;
-    let layer = avm.resolve_layer(level_id, context);
-    let target = layer.object().as_object()?;
-
-    let method = args.get(2).cloned().unwrap_or(Value::Undefined);
-    let method = NavigationMethod::from_method_str(&method.coerce_to_string(avm, context)?);
-    let (url, opts) = avm.locals_into_request_options(context, url, method);
-    let fetch = context.navigator.fetch(url, opts);
-    let process =
-        context
-            .load_manager
-            .load_form_into_object(context.player.clone().unwrap(), target, fetch);
-
-    context.navigator.spawn_future(process);
-
-    Ok(Value::Undefined.into())
-}
-
-pub fn unload_movie<'gc>(
-    avm: &mut Avm1<'gc>,
-    context: &mut UpdateContext<'_, 'gc, '_>,
-    _this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<ReturnValue<'gc>, Error> {
-    let target_or_path = args.get(0).cloned().unwrap_or(Value::Undefined);
-    let target = if let Value::Object(target) = target_or_path {
-        let target = target.as_display_object();
-
-        if target.is_none() {
-            log::warn!("unloadMovie: Object passed as target is not a movie clip");
-        }
-
-        target
-    } else {
-        let target_path = target_or_path.into_string();
-        let target_clip =
-            Avm1::resolve_dot_path_clip(avm.target_clip(), context.root, &target_path);
-
-        if target_clip.is_none() {
-            log::warn!("unloadMovie: Could not resolve target path {}", target_path)
-        }
-
-        target_clip
-    };
-
-    if let Some(mut target) = target {
-        //TODO: How do we reclaim resources from unloaded movies?
-        if let Some(parent) = target.parent() {
-            parent
-                .as_movie_clip()
-                .unwrap()
-                .remove_child_from_avm(context, target);
-        } else {
-            //Removing a layer is a little different.
-            let mut layer_depth = None;
-
-            for (depth, layer) in context.layers.iter() {
-                if DisplayObject::ptr_eq(*layer, target) {
-                    layer_depth = Some(*depth);
-                    break;
-                }
-            }
-
-            if let Some(depth) = layer_depth {
-                context.layers.remove(&depth);
-            }
-
-            target.unload(context);
-        }
-    }
-
-    Ok(Value::Undefined.into())
-}
-
-pub fn unload_movie_num<'gc>(
-    avm: &mut Avm1<'gc>,
-    context: &mut UpdateContext<'_, 'gc, '_>,
-    _this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<ReturnValue<'gc>, Error> {
-    let level_id = args
-        .get(1)
-        .cloned()
-        .unwrap_or(Value::Undefined)
-        .as_number(avm, context)? as u32;
-
-    if let Some(mut layer) = context.layers.get_mut(&level_id).copied() {
-        layer.unload(context);
-    }
-
-    context.layers.remove(&level_id);
-
-    Ok(Value::Undefined.into())
 }
 
 /// This structure represents all system builtins that are used regardless of
@@ -552,48 +329,6 @@ pub fn create_globals<'gc>(
     globals.force_set_function(
         "ASSetPropFlags",
         object::as_set_prop_flags,
-        gc_context,
-        EnumSet::empty(),
-        Some(function_proto),
-    );
-    globals.force_set_function(
-        "loadMovie",
-        load_movie,
-        gc_context,
-        EnumSet::empty(),
-        Some(function_proto),
-    );
-    globals.force_set_function(
-        "loadMovieNum",
-        load_movie_num,
-        gc_context,
-        EnumSet::empty(),
-        Some(function_proto),
-    );
-    globals.force_set_function(
-        "loadVariables",
-        load_variables,
-        gc_context,
-        EnumSet::empty(),
-        Some(function_proto),
-    );
-    globals.force_set_function(
-        "loadVariablesNum",
-        load_variables_num,
-        gc_context,
-        EnumSet::empty(),
-        Some(function_proto),
-    );
-    globals.force_set_function(
-        "unloadMovie",
-        unload_movie,
-        gc_context,
-        EnumSet::empty(),
-        Some(function_proto),
-    );
-    globals.force_set_function(
-        "unloadMovieNum",
-        unload_movie_num,
         gc_context,
         EnumSet::empty(),
         Some(function_proto),
